feat(condo): DOMA-8258 refactoring for PR comments
