+++ apps/condo/domains/notification/constants/links.js
@@ -1,5 +1,7 @@
-const UNSUBSCRIBE_LINK_PATH = '/unsubscribe'
+const UNSUBSCRIBE_LINK_PATH = '/api/unsubscribe'
+const UNSUBSCRIBED_PAGE_PATH = '/unsubscribed'
 
 module.exports = {
     UNSUBSCRIBE_LINK_PATH,
+    UNSUBSCRIBED_PAGE_PATH,
 }
