+++ apps/condo/lang/en/messages/BILLING_RECEIPT_FILE_ADDED/email.html.njk
@@ -18,6 +18,6 @@
 <br>
 <br>
 If you do not want to receive emails with receipts, please click <a
-        href="{{ env.serverUrl }}/unsubscribe?id={{ message.id }}">unsubscribe</a>
+        href="{{ env.serverUrl }}/api/unsubscribe?id={{ message.id }}">unsubscribe</a>
 </body>
 </html>
