+++ apps/condo/lang/ru/messages/BILLING_RECEIPT_FILE_ADDED/email.html.njk
@@ -20,6 +20,6 @@
 <br>
 <br>
 Если вы не хотите получать письма с квитанциями, перейдите по ссылке <a
-        href="{{ env.serverUrl }}/unsubscribe?id={{ message.id }}">отписаться</a>
+        href="{{ env.serverUrl }}/api/unsubscribe?id={{ message.id }}">отписаться</a>
 </body>
 </html>
